// clients/js/src/generated/types/basePluginAdapterKey.ts
/**
 * This code was AUTOGENERATED using the kinobi library.
 * Please DO NOT EDIT THIS FILE, instead use visitors
 * to add features, then rerun kinobi to update it.
 *
 * @see https://github.com/metaplex-foundation/kinobi
 */

import { PublicKey } from '@metaplex-foundation/umi';
import {
  GetDataEnumKind,
  GetDataEnumKindContent,
  Serializer,
  dataEnum,
  publicKey as publicKeySerializer,
  struct,
  tuple,
} from '@metaplex-foundation/umi/serializers';
import {
  BasePluginAuthority,
  BasePluginAuthorityArgs,
  getBasePluginAuthoritySerializer,
} from '.';

export type BasePluginAdapterKey =
  | { __kind: 'LifecycleHook'; fields: [PublicKey] }
  | { __kind: 'Oracle'; fields: [PublicKey] }
  | { __kind: 'DataStore'; fields: [BasePluginAuthority] };

export type BasePluginAdapterKeyArgs =
  | { __kind: 'LifecycleHook'; fields: [PublicKey] }
  | { __kind: 'Oracle'; fields: [PublicKey] }
  | { __kind: 'DataStore'; fields: [BasePluginAuthorityArgs] };

export function getBasePluginAdapterKeySerializer(): Serializer<
  BasePluginAdapterKeyArgs,
  BasePluginAdapterKey
> {
  return dataEnum<BasePluginAdapterKey>(
    [
      [
        'LifecycleHook',
        struct<GetDataEnumKindContent<BasePluginAdapterKey, 'LifecycleHook'>>([
          ['fields', tuple([publicKeySerializer()])],
        ]),
      ],
      [
        'Oracle',
        struct<GetDataEnumKindContent<BasePluginAdapterKey, 'Oracle'>>([
          ['fields', tuple([publicKeySerializer()])],
        ]),
      ],
      [
        'DataStore',
        struct<GetDataEnumKindContent<BasePluginAdapterKey, 'DataStore'>>([
          ['fields', tuple([getBasePluginAuthoritySerializer()])],
        ]),
      ],
    ],
    { description: 'BasePluginAdapterKey' }
  ) as Serializer<BasePluginAdapterKeyArgs, BasePluginAdapterKey>;
}

// Data Enum Helpers.
export function basePluginAdapterKey(
  kind: 'LifecycleHook',
  data: GetDataEnumKindContent<
    BasePluginAdapterKeyArgs,
    'LifecycleHook'
  >['fields']
): GetDataEnumKind<BasePluginAdapterKeyArgs, 'LifecycleHook'>;
export function basePluginAdapterKey(
  kind: 'Oracle',
  data: GetDataEnumKindContent<BasePluginAdapterKeyArgs, 'Oracle'>['fields']
): GetDataEnumKind<BasePluginAdapterKeyArgs, 'Oracle'>;
export function basePluginAdapterKey(
  kind: 'DataStore',
  data: GetDataEnumKindContent<BasePluginAdapterKeyArgs, 'DataStore'>['fields']
): GetDataEnumKind<BasePluginAdapterKeyArgs, 'DataStore'>;
export function basePluginAdapterKey<
  K extends BasePluginAdapterKeyArgs['__kind'],
>(kind: K, data?: any): Extract<BasePluginAdapterKeyArgs, { __kind: K }> {
  return Array.isArray(data)
    ? { __kind: kind, fields: data }
    : { __kind: kind, ...(data ?? {}) };
}
export function isBasePluginAdapterKey<
  K extends BasePluginAdapterKey['__kind'],
>(
  kind: K,
  value: BasePluginAdapterKey
): value is BasePluginAdapterKey & { __kind: K } {
  return value.__kind === kind;
}