// clients/js/src/generated/types/adapterRegistryRecord.ts
/**
 * This code was AUTOGENERATED using the kinobi library.
 * Please DO NOT EDIT THIS FILE, instead use visitors
 * to add features, then rerun kinobi to update it.
 *
 * @see https://github.com/metaplex-foundation/kinobi
 */

import { Option, OptionOrNullable } from '@metaplex-foundation/umi';
import {
  Serializer,
  array,
  option,
  struct,
  tuple,
  u64,
} from '@metaplex-foundation/umi/serializers';
import {
  AdapterCheckResult,
  AdapterCheckResultArgs,
  BasePluginAuthority,
  BasePluginAuthorityArgs,
  HookableLifecycleEvent,
  HookableLifecycleEventArgs,
  PluginAdapterType,
  PluginAdapterTypeArgs,
  getAdapterCheckResultSerializer,
  getBasePluginAuthoritySerializer,
  getHookableLifecycleEventSerializer,
  getPluginAdapterTypeSerializer,
} from '.';

export type AdapterRegistryRecord = {
  pluginType: PluginAdapterType;
  authority: BasePluginAuthority;
  lifecycleChecks: Option<Array<[HookableLifecycleEvent, AdapterCheckResult]>>;
  offset: bigint;
  dataOffset: Option<bigint>;
  dataLen: Option<bigint>;
};

export type AdapterRegistryRecordArgs = {
  pluginType: PluginAdapterTypeArgs;
  authority: BasePluginAuthorityArgs;
  lifecycleChecks: OptionOrNullable<
    Array<[HookableLifecycleEventArgs, AdapterCheckResultArgs]>
  >;
  offset: number | bigint;
  dataOffset: OptionOrNullable<number | bigint>;
  dataLen: OptionOrNullable<number | bigint>;
};

export function getAdapterRegistryRecordSerializer(): Serializer<
  AdapterRegistryRecordArgs,
  AdapterRegistryRecord
> {
  return struct<AdapterRegistryRecord>(
    [
      ['pluginType', getPluginAdapterTypeSerializer()],
      ['authority', getBasePluginAuthoritySerializer()],
      [
        'lifecycleChecks',
        option(
          array(
            tuple([
              getHookableLifecycleEventSerializer(),
              getAdapterCheckResultSerializer(),
            ])
          )
        ),
      ],
      ['offset', u64()],
      ['dataOffset', option(u64())],
      ['dataLen', option(u64())],
    ],
    { description: 'AdapterRegistryRecord' }
  ) as Serializer<AdapterRegistryRecordArgs, AdapterRegistryRecord>;
}